requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Werkzeug==2.3.7
Pillow==11.2.1
python-dotenv==1.0.0 